requirements.txt
google-genai
requests
selenium
mysql-connector-python
python-dotenv
scrapy
pypdf